Add validation of sort and filter parameters
This is a little more involved than it might otherwise be since these parameters in part
are directly inserted into a SQL query string. Those that are so inserted are strictly
white listed. Also in this commit is the Exception mapper for giving useful information
to the user when an parameter does not pass validation.